Data_Processing/Anomaly_Validator/requirements.txt
# Anomaly Validator - LLM Multi-Agent Framework
# Dependencies for Paper 2 implementation

# Core LLM
google-generativeai>=0.3.0  # Gemini API

# Document Processing
PyPDF2>=3.0.0               # PDF parsing
pdfplumber>=0.10.0          # Advanced PDF extraction
openpyxl>=3.1.0             # Excel files
python-docx>=1.1.0          # Word documents
pytesseract>=0.3.10         # OCR for images
Pillow>=10.0.0              # Image processing

# Web Scraping
beautifulsoup4>=4.12.0      # HTML parsing
selenium>=4.15.0            # Dynamic page scraping
newspaper3k>=0.2.8          # News article extraction
requests>=2.31.0            # HTTP requests
aiohttp>=3.9.0              # Async requests
lxml>=4.9.0                 # XML/HTML parsing

# Data Processing
pandas>=2.0.0               # Data manipulation
numpy>=1.24.0               # Numerical operations

# Configuration
PyYAML>=6.0                 # YAML config files
python-dotenv>=1.0.0        # Environment variables

# Utilities
tenacity>=8.2.0             # Retry logic
tqdm>=4.66.0                # Progress bars
colorama>=0.4.6             # Colored terminal output

